+++ borax/__init__.py
@@ -1,1 +1,8 @@
-from . import tagging, bibtex_exporter, metadata_fetcher, history_tracker, utils, library_config
+from . import (
+    tagging,
+    bibtex_exporter,
+    metadata_fetcher,
+    history_tracker,
+    utils,
+    library_config,
+)